backend/requirements.txt
# Web Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]==0.30.0

# Auth & JWT
python-jose[cryptography]==3.3.0
httpx==0.27.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.0

# Cache & Sessions
redis==5.0.4

# File Storage (works with MinIO locally + AWS S3 later - same API)
boto3==1.34.0

# Data Validation
pydantic==2.8.2
pydantic-settings==2.3.0
python-multipart==0.0.9

# MCP SDK
mcp==1.0.0

# Utilities
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
